backend.service.chore: refactor parcel check in service into partial class
